chore(otelcol): update Docker image & add Jaeger
This change upgrades the version of the OpenTelemetry collector to the
vanilla Docker image (as opposed to the *-contrib image) and adds a
running Jaeger container to the supporting services in Docker Compose.

Prior to this change our `otelcol` container had a configuration that
pointed all incoming `otlp` traffic at the Honeycomb API for ingestion
and analysis. While this will likely be our path when running in
production, Jaeger is not a bad option for us in development for one
important reason: no sampling. This updated configuration does do
batching but does *not* sample any tracing data passing through it. This
allows us as developers/operators to see and track any errors we come
across when running our software. A sampling approach might omit the
very trace data we are looking. And from past experience, our team knows
that it's rude to dump an excessive amount of tracing info to another
company's API endpoints (which was our default behavior).

Currently, the Jaeger instance is the "all-in-one" and is listening for
incoming OTLP/gRPC traffic. The UI front-end is accessible at
http://localhost:16686.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>