chore(pinga,dal,nats-subscriber): add initial job telemetry in Pinga
This changes enables OpenTelemetry exporting from the Pinga service by
default (not that there is a `--disable-opentelemetry` flag on all
services) and attempts to better instrument the core tasks in Pinga:
executing a job.

This change adds a more runtime metadata useful for telemetry such as a
unique`JobInvocationId` for every job that Pinga runs, a unique instance
ID for each running instance of the service as well as error reporting
and the closing of the job execution span with an explicit `status = OK`
or `status = ERROR`.

As a bonus, the spawned job execution Tokio tasks are now named so that
we might be able to better track Pinga running in tokio-console. Cool,
no?

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>